+++ src/data/extension/vec_types.rs
@@ -1,5 +1,5 @@
 pub(crate) trait ChopLineBreak {
-    /// Removes trailing `\n` or `\r\n`
+    /// Removes a trailing `\n` or `\r\n`
     fn chop_line_break(&mut self);
 }
 
@@ -13,5 +13,25 @@ impl ChopLineBreak for Vec<u8> {
                 self.pop();
             }
         }
+    }
+}
+
+pub(crate) trait StripLineBreak {
+    /// Returns a new slice without a trailing `\n` or `\r\n`
+    fn strip_line_break(&self) -> Self;
+}
+
+impl<'a> StripLineBreak for &'a [u8] {
+    #[inline]
+    fn strip_line_break(&self) -> &'a [u8] {
+        let mut out = *self;
+        if out.ends_with(b"\n") {
+            out = &out[..out.len() - 1];
+
+            if out.ends_with(b"\r") {
+                out = &out[..out.len() - 1];
+            }
+        }
+        out
     }
 }