+++ src/data/records/fasta/test.rs
@@ -2,6 +2,17 @@ use super::*;
 use std::io::Cursor;
 
 #[test]
+fn single_seq() {
+    let mut reader = FastaReader::new(Cursor::new(">seq1\nA"));
+
+    let record1 = reader.next().unwrap().unwrap();
+    assert_eq!(record1.name, "seq1");
+    assert_eq!(record1.sequence, b"A");
+
+    assert!(reader.next().is_none());
+}
+
+#[test]
 fn sequence_whitespace() {
     let mut reader = FastaReader::new(Cursor::new(">seq1\nATG C   \n>seq2\r\n\n  AC G\r\n T"));
 
@@ -20,7 +31,7 @@ fn sequence_whitespace() {
 fn empty_file() {
     let mut reader = FastaReader::new(Cursor::new(""));
 
-    assert_eq!(reader.next().unwrap().unwrap_err().to_string(), "No FASTA data found.");
+    assert_eq!(reader.next().unwrap().unwrap_err().to_string(), "No FASTA data found!");
 
     // Ensure iterator terminates
     assert!(reader.count() < 100);
@@ -30,7 +41,7 @@ fn empty_file() {
 fn whitespace_only() {
     let mut reader = FastaReader::new(Cursor::new("   \r\n \r\t\n   \t"));
 
-    assert_eq!(reader.next().unwrap().unwrap_err().to_string(), "No FASTA data found.");
+    assert_eq!(reader.next().unwrap().unwrap_err().to_string(), "No FASTA data found!");
 
     // Ensure iterator terminates
     assert!(reader.count() < 100);
@@ -63,7 +74,10 @@ fn empty_header_first_record() {
 fn empty_sequence_first_record() {
     let mut reader = FastaReader::new(Cursor::new(">seq1\n"));
 
-    assert_eq!(reader.next().unwrap().unwrap_err().to_string(), "Missing FASTA sequence!");
+    assert_eq!(
+        reader.next().unwrap().unwrap_err().to_string(),
+        "Missing FASTA sequence! See header: seq1"
+    );
 
     // Ensure iterator terminates
     assert!(reader.count() < 100);
@@ -90,7 +104,10 @@ fn empty_sequence_second_record() {
     assert_eq!(record1.name, "seq1");
     assert_eq!(record1.sequence, b"GCAT");
 
-    assert_eq!(reader.next().unwrap().unwrap_err().to_string(), "Missing FASTA sequence!");
+    assert_eq!(
+        reader.next().unwrap().unwrap_err().to_string(),
+        "Missing FASTA sequence! See header: seq2"
+    );
 
     // Ensure iterator terminates
     assert!(reader.count() < 100);
@@ -102,7 +119,7 @@ fn invalid_char_header_first_record() {
 
     assert_eq!(
         reader.next().unwrap().unwrap_err().to_string(),
-        "FASTA records must start with the '>' symbol on a newline, and no other '>' symbols can occur in a header!"
+        "FASTA records must start with the '>' symbol on a newline, and no other '>' symbols can occur in a header! See header: seq1>seq2"
     );
 
     // Ensure iterator terminates
@@ -113,12 +130,9 @@ fn invalid_char_header_first_record() {
 fn invalid_char_sequence_first_record() {
     let mut reader = FastaReader::new(Cursor::new(">seq1\r\nGADGSDHS>CAT\n>seq2\nGAT"));
 
-    let record1 = reader.next().unwrap().unwrap();
-    assert_eq!(record1.name, "seq1");
-    assert_eq!(record1.sequence, b"GADGSDHS");
     assert_eq!(
         reader.next().unwrap().unwrap_err().to_string(),
-        "FASTA records must start with the '>' symbol on a newline, and no other '>' symbols can occur in a sequence!"
+        "FASTA records must start with the '>' symbol on a newline, and no other '>' symbols can occur in a sequence! See header: seq1"
     );
 
     // Ensure iterator terminates
@@ -134,7 +148,7 @@ fn invalid_char_header_second_record() {
     assert_eq!(record1.sequence, b"GADGSDHS");
     assert_eq!(
         reader.next().unwrap().unwrap_err().to_string(),
-        "FASTA records must start with the '>' symbol on a newline, and no other '>' symbols can occur in a header!"
+        "FASTA records must start with the '>' symbol on a newline, and no other '>' symbols can occur in a header! See header: seq2>seq3"
     );
 
     // Ensure iterator terminates
@@ -148,7 +162,7 @@ fn invalid_char_header_second_record() {
 
     assert_eq!(
         reader.next().unwrap().unwrap_err().to_string(),
-        "FASTA records must start with the '>' symbol on a newline, and no other '>' symbols can occur in a header!"
+        "FASTA records must start with the '>' symbol on a newline, and no other '>' symbols can occur in a header! See header: seq2>invalid"
     );
 
     // Ensure iterator terminates
@@ -165,15 +179,9 @@ fn invalid_char_sequence_second_record() {
     assert_eq!(name, "seq1");
     assert_eq!(sequence, b"GADGSDHS");
 
-    let Some(Ok(FastaSeq { name, sequence })) = reader.next() else {
-        panic!("The first record should process without error.");
-    };
-    assert_eq!(name, "seq2");
-    assert_eq!(sequence, b"GAT");
-
     assert_eq!(
         reader.next().unwrap().unwrap_err().to_string(),
-        "FASTA records must start with the '>' symbol on a newline, and no other '>' symbols can occur in a sequence!"
+        "FASTA records must start with the '>' symbol on a newline, and no other '>' symbols can occur in a sequence! See header: seq2"
     );
 
     // Ensure iterator terminates